Add files via upload
There is lot of uncertainties on demand of electricity in the global market. Building a model to predict 
the future electricity demand would be very useful. There are various factors which are needed to be 
considered for the prediction of the demand, some of them are climate, socio-economic activities , 
population etc. In the model we are going to see different forecasting methods used for predicting the 
demand.
In the model we are doing comparative and testing analysis between different parameters. We try to 
reduce the error percentage to get an accurate reading .